% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fpkm_to_tpm}
\alias{fpkm_to_tpm}
\title{fpkm_to_tpm}
\usage{
fpkm_to_tpm(fpkm)
}
\arguments{
\item{fpkm}{gene by sample matrix in FPKM.}
}
\value{
gene by sample matrix in TPM.
}
\description{
convert FPKM to TPM
}
\details{
Convert Fragments Per Kilobase of transcript per Million mapped reads (FPKM) to Transcripts Per Million (TPM).
}
