% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LymphoMapR-package.R
\name{LymphoMapR-package}
\alias{LymphoMapR-package}
\title{LymphoMapR package}
\description{
LymphoMapR: Lymphoma Microenvironment Archetype Profile (LymphoMAP) classifer in R
}
\references{
Xubin Li, Kartik Singhal, Qing Deng, Dai Chihara, David A Russler-Germain, Usama Hussein, Jennifer Ann Foltz, Jared Henderson, Ashley Wilson, Evelyn Schmidt, Imran A. Nizamuddin, Tommy Dinh, Ryan Sun, Akhil Kesaraju, Laura K. Hilton, David W. Scott, Francisco Vega, Christopher R. Flowers, Obi Griffith, Todd A Fehniger, Malachi Griffith, Michael R. Green,
Comprehensive Characterization and Validation of the Tumor Microenvironment in Patients with Relapsed/Refractory Large B-Cell Lymphoma Identifies Subgroups with Greatest Benefit from CD19 CAR T-Cell Therapy,
Blood, Volume 144, Supplement 1, 2024. https://doi.org/10.1182/blood-2024-204166.
}
\keyword{package}
