% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_confmat}
\alias{plot_confmat}
\title{plot_confmat}
\usage{
plot_confmat(confmat)
}
\arguments{
\item{confmat}{a confusion matrix object returned by caret::confusionMatrix().}
}
\value{
a Heatmap object returned by ComplexHeatmap::draw().
}
\description{
Plot heatmap of confusion matrix
}
\details{
Plot heatmap of confusion matrix returned by caret::confusionMatrix().
}
