% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_umap}
\alias{plot_umap}
\title{plot_umap}
\usage{
plot_umap(mt, sample_info, npcs = NULL, palette = NULL)
}
\arguments{
\item{mt}{sample by gene expression matrix.}

\item{sample_info}{sample categories to be shown on the scatter plot.}

\item{npcs}{number of principal components used for UMAP calculation; if NULL, calculate UMAP using the input matrix without PCA. Default = NULL.}

\item{palette}{palette of sample categories. Default = NULL}
}
\value{
a ggplot object.
}
\description{
Plot UMAP
}
\details{
Calculate and generate a UMAP plot.
}
