% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_umap_archetype}
\alias{plot_umap_archetype}
\title{plot_umap_archetype}
\usage{
plot_umap_archetype(mt, sample_archetypes, npcs = NULL)
}
\arguments{
\item{mt}{sample by gene expression matrix.}

\item{sample_archetypes}{sample archetypes.}

\item{npcs}{number of principal components used for UMAP calculation; if NULL, calculate UMAP using the input matrix without PCA. Default = NULL.}
}
\value{
a ggplot object.
}
\description{
Plot UMAP by archetypes
}
\details{
Calculate and generate a UMAP plot showing archetypes.
}
