% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.LymphoMapObject}
\alias{print.LymphoMapObject}
\title{Print method for LymphoMapObject}
\usage{
\method{print}{LymphoMapObject}(res, ...)
}
\arguments{
\item{res}{A \code{LymphoMapObject}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\description{
Print method for LymphoMapObject
}
