% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_LymphoMap}
\alias{run_LymphoMap}
\title{run_LymphoMap}
\usage{
run_LymphoMap(
  expr_pred,
  data_type = c("RNAseq_TPM_log2p1", "microarray_lognorm", "RNAseq_counts",
    "NanoString_norm_counts"),
  LymphoMAP_prob_cutoff = 0,
  n_features_per_archetype = 50,
  min_fract_poszscore_group = 0.5,
  max_fract_poszscore_rest = 0.5,
  min_log2FoldChange = 0.58,
  padj_cutoff = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{expr_pred}{gene by sample expression matrix of prediction data. Row names: gene symbols, column names: sample names.}

\item{data_type}{expression data type from "RNAseq_TPM_log2p1", "microarray_lognorm", "RNAseq_counts", "NanoString_norm_counts". Default = "RNAseq_TPM_log2p1".}

\item{LymphoMAP_prob_cutoff}{LymphoMAP assignment threshold of classification probability. Default = 0.}

\item{n_features_per_archetype}{number of features per archetype used in model training. Default = 50.}

\item{min_fract_poszscore_group}{criterion for selected features that have minimum fraction of samples within archetype. Default = 0.5.}

\item{max_fract_poszscore_rest}{criterion for selected features that have maximum fraction of samples outside archetype. Default = 0.5.}

\item{min_log2FoldChange}{criterion for selected features that have minimum log2FoldChange of archetype vs rest. Default = 0.58.}

\item{padj_cutoff}{criterion for selected features that have maximum padj of archetype vs rest. Default = 0.05.}

\item{verbose}{logical value if showing all messages. Default = FALSE.}
}
\value{
A named list of class `"LymphoMapObject"` with the following components:
\describe{
  \item{data}{A named list containing:
    \describe{
      \item{\code{X_train}}{Training expression matrix (numeric).}
      \item{\code{y_train}}{Training labels (named factor).}
      \item{\code{X_pred}}{Prediction expression matrix (numeric).}
      \item{\code{feature}}{Selected features (tibble).}
      \item{\code{feature_archetypes}}{Archetypes of selected features (named factor).}
    }
  }
  \item{\code{nb_model}}{An object of class \code{"naive_bayes"} containing the trained model.}
  \item{\code{cv_confmat}}{A confusion matrix object returned by \code{caret::confusionMatrix()}.}
  \item{\code{pred}}{Final LymphoMAP label assignments and probabilities. The last three columns contain the classification probability matrix for all archetypes. (tibble).}
  \item{\code{y_pred}}{Predicted labels (named factor).}
  \item{\code{y_prob}}{Probabilities of predicted labels (named numeric vector).}
  \item{\code{parameters}}{A list of parameters used in the function call.}
}
}
\description{
Run LymphoMAP classification
}
\details{
run_LymphoMap integrates prediction expression matrix with training data, selects archetype-defining features from common genes shared by both datasets, fits a naive Bayes model, and finally assign LymphoMAP labels based on classification probabilities for prediction samples.
}
